hw5.Work on bugs 1

+++ homework5/App/Controllers/AdminPanel/Delete.php
@@ -8,14 +8,17 @@ class Delete extends AdminPanelController
 {
 
     /**
-     * @throws \App\Exceptions\Db
-     * @throws \App\Exceptions\DbNotFoundRecord
+     * @throws \App\Exceptions\DbException
+     * @throws \App\Exceptions\E404Exception
      */
     protected function action()
     {
         if ( isset( $_GET['id'] ) ) {
             $article = \App\Models\Article::findById( $_GET['id'] ); //Article
-            $article->delete();
+
+            if (false !== $article) {
+                $article->delete();
+            }
         }
 
         header('Location: /homework5/adminPanel/index/');
